add: c_handler realisation
