--- polygon4.py
@@ -1,56 +1,0 @@
-from collections.abc import Iterator
-from typing import Any
-from pysatl_tsp.core import Handler
-from pysatl_tsp._c.lib import *
-import cffi
-
-ffi = cffi.FFI()
-
-
-class CPIHandler(Handler[float, float]):
-
-    def __init__(self, source: Handler[Any, float] | None = None):
-        super().__init__(source)
-        if source is not None:
-            self.handler = createHandler(ffi.NULL, source.handler, sumdiv, ffi.NULL)
-        else:
-            self.handler = createHandler(ffi.NULL, ffi.NULL, sumdiv, ffi.NULL)
-
-    def __iter__(self) -> Iterator[float | None]:
-        if self.source is None:
-            raise ValueError("Source is not set")
-        self.src_itr = iter(self.source)
-        self.handler.module = ffi.cast("void*", id(self.src_itr))
-        return self
-
-    def __next__(self):
-        length = ffi.new("int [1]")
-        res = applyIter(self.handler, length)
-        if res != ffi.NULL:
-            tmp = ffi.unpack(res, length[0])
-            ffi.gc(res, free)
-            return tmp[0]
-        else:
-            raise StopIteration
-
-    def __del__(self):
-        freeHandler(self.handler)
-
-
-
-from pysatl_tsp.core.data_providers import CSimpleDataProvider
-
-data = [1.5]
-for i in range(1, 10):
-    data.append(data[0] + i)
-print(data)
-provider = CSimpleDataProvider(data)
-for elem in provider:
-    print(elem, end="  ")
-print("\n-----------------------")
-pipeline = (provider | CPIHandler())
-for elem in pipeline:
-    print(elem, end=" ")
-print("\n")
-for elem in CPIHandler(provider):
-    print(elem)
