+++ pysatl_tsp/__init__.py
@@ -1,1 +1,1 @@
-#from ._c.lib import *
+from ._c.lib import *
