+++ pysatl_tsp/core/data_providers/__init__.py
@@ -3,7 +3,6 @@ from .database_data_provider import DatabaseAdapter, DataBaseDataProvider
 from .file_data_provider import FileDataProvider
 from .simple_data_provider import SimpleDataProvider
 from .websocket_data_provider import WebSocketDataProvider
-from .c_simple_data_provider import CSimpleDataProvider
 
 __all__ = [
     "DataBaseDataProvider",
@@ -13,5 +12,4 @@ __all__ = [
     "SimpleDataProvider",
     "T",
     "WebSocketDataProvider",
-    "CSimpleDataProvider",
 ]
