+++ pysatl_tsp/core/handler.py
@@ -111,7 +111,9 @@ class Pipeline(Handler[T, V]):
         if hasattr(self.second, 'handler'):
             if hasattr(self.first, 'handler'):
                 self.second.handler.source = self.first.handler
-                self.handler = self.second.handler
+            else:
+                self.second.handler.source = ffi.NULL
+            self.handler = self.second.handler
 
     def __iter__(self) -> Iterator[V]:
         """Create an iterator that processes data through both handlers in sequence.
